+++ sapy_modules/gtk/Add_mom_handler.py
@@ -1,0 +1,6 @@
+import gi
+gi.require_version('Gtk', '3.0')
+from gi.repository import Gtk
+
+class AddMomHandler:
+    pass
